fastapi tutorial: added authentication, users

+++ fastapi-tutorial/app/app.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
-from app.schemas import CreatePost, PostResponse
-from app.db import Post, create_db_and_tables, get_async_session
+from app.schemas import CreatePost, PostResponse, UserRead, UserCreate, UserUpdate
+from app.db import Post, create_db_and_tables, get_async_session, User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy import select
@@ -10,6 +10,8 @@ import shutil
 import os
 import uuid
 import tempfile
+from app.users import auth_backend, current_active_user, fastapi_users
+
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
@@ -18,9 +20,34 @@ async def lifespan(app: FastAPI):
 
 app = FastAPI(lifespan=lifespan)
 
+app.include_router(
+    fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"]
+)
+app.include_router(
+    fastapi_users.get_register_router(UserRead, UserCreate),
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_reset_password_router(),
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_verify_router(UserRead),
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_users_router(UserRead, UserUpdate),
+    prefix="/users",
+    tags=["users"],
+)
+
 @app.post("/upload")
 async def upload_file(
     file: UploadFile = File(...),
+    user: User = Depends(current_active_user),
     caption: str = Form(""),
     session: AsyncSession = Depends(get_async_session),
 ):
@@ -41,7 +68,9 @@ async def upload_file(
         )
 
         if upload_result.response_metadata.http_status_code == 200:
-            post = Post(caption = caption, url = upload_result.url, 
+            post = Post(caption = caption, 
+                        url = upload_result.url, 
+                        user_id = user.id,
                         file_type = "video" if file.content_type.startswith("video/") else "image",
                         file_name = upload_result.name)
             session.add(post)
@@ -57,27 +86,34 @@ async def upload_file(
 
 
 @app.get("/feed")
-async def get_feed(session: AsyncSession = Depends(get_async_session)):
+async def get_feed(session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
     data = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in data.all()]
+
+    result = await session.execute(select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id: u.email for u in users}
 
     posts_data=[]
     for post in posts:
         posts_data.append(
             {
                 "id": str(post.id),
+                "user_id": str(post.user_id),
                 "caption": post.caption,
                 "url": post.url,
                 "file_type": post.file_type,
                 "file_name": post.file_name,
                 "created_at": post.created_at.isoformat(),
+                "is_owner": post.user_id == user.id,
+                "email": user_dict[post.user_id],
             }
     )
     
     return {"posts": posts_data}
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_session), user: User = Depends(current_active_user)):
     try:
         post_uuid= uuid.UUID(post_id)
         post = await session.execute(select(Post).where(Post.id == post_uuid))
@@ -86,6 +122,9 @@ async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_ses
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
         
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="You are not authorized to delete this post")
+
         await session.delete(post)
         await session.commit()
         return {"message": "Post deleted successfully"}
